+++ config.py
@@ -19,23 +19,30 @@ class Config:
         "ready": "py_dummy_service.healthz.readiness",
     }
 
+    @staticmethod
+    def init_app(app):
+        pass
+
 
 class DevelopmentConfig(Config):
     DEBUG = True
     TESTING = False
     DEBUG_METRICS = 1
+    LOG_LEVEL = "DEBUG"
 
 
 class TestingConfig(Config):
     DEBUG = False
     TESTING = True
     DEBUG_METRICS = 0
+    LOG_LEVEL = "INFO"
 
 
 class ProductionConfig(Config):
     DEBUG = False
     TESTING = False
     DEBUG_METRICS = 0
+    LOG_LEVEL = "INFO"
 
 
 config = {